Playing with news posts - I don't like it

+++ app/concepts/post/cell/show.rb
@@ -18,14 +18,18 @@ module Post::Cell
     end
 
     def edit
-      if tyrant.current_user.id.to_s == model.content["user_id"] or tyrant.current_user.email == "admin@email.com" #change in order to have policy
-        link_to "Edit", edit_post_path(model.id)
+      unless tyrant.current_user == nil
+        if tyrant.current_user.id.to_s == model.content["user_id"] or tyrant.current_user.email == "admin@email.com" #change in order to have policy
+          link_to "Edit", edit_post_path(model.id)
+        end
       end
     end
 
     def delete
-      if tyrant.current_user.id.to_s == model.content["user_id"] or tyrant.current_user.email == "admin@email.com" #change in order to have policy
-        link_to "Delete", post_path(model.id), method: :delete, data: {confirm: 'Are you sure?'}
+      unless tyrant.current_user == nil
+        if tyrant.current_user.id.to_s == model.content["user_id"] or tyrant.current_user.email == "admin@email.com" #change in order to have policy
+          link_to "Delete", post_path(model.id), method: :delete, data: {confirm: 'Are you sure?'}
+        end
       end
     end
